requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==22.1.0
billiard==3.6.4.0
boto3==1.24.71
botocore==1.27.96
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
Django==4.0.7
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-redis==5.2.0
django-storages==1.13.1
django-timezone-field==6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-jwt==1.19.2
drf-spectacular==0.24.2
drf-spectacular-sidecar==2023.9.1
factory-boy==3.2.1
Faker==19.6.1
gunicorn==20.1.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.17.3
kombu==5.3.2
packaging==23.1
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.8.0
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
s3transfer==0.6.2
sentry-sdk==1.9.8
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
